Refactor contexts to folder
