+++ app/src/components/Chart.tsx
@@ -24,13 +24,13 @@ import {
 } from "../const"
 import { unionAuthors } from "../util"
 import { Legend } from "./Legend"
-import { ChartType, useOptions } from "../OptionsContext"
+import { ChartType, useOptions } from "../contexts/OptionsContext"
 import styled from "styled-components"
 import { Tooltip } from "./Tooltip"
-import { useSearch } from "../SearchContext"
-import { useData } from "./DataContext"
+import { useSearch } from "../contexts/SearchContext"
+import { useData } from "../contexts/DataContext"
 import { animated, useSpring } from "@react-spring/web"
-import { useMetricCaches } from "../MetricContext"
+import { useMetricCaches } from "../contexts/MetricContext"
 
 type CircleOrRectHiearchyNode =
   | HierarchyCircularNode<HydratedGitObject>
