+++ app/src/components/Options.tsx
@@ -1,7 +1,7 @@
 import { Metric, MetricType } from "../metrics"
 import { Box } from "./util"
 import { EnumSelect } from "./EnumSelect"
-import { Chart, ChartType, useOptions } from "../OptionsContext"
+import { Chart, ChartType, useOptions } from "../contexts/OptionsContext"
 import { Spacer } from "./Spacer"
 
 export function Options() {
