+++ app/src/components/Providers.tsx
@@ -10,11 +10,11 @@ import {
   getDefaultOptions,
   Options,
   OptionsContext,
-} from "../OptionsContext"
-import { SearchContext } from "./../SearchContext"
+} from "../contexts/OptionsContext"
+import { SearchContext } from "../contexts/SearchContext"
 import { HydratedGitBlobObject, ParserData } from "../../../parser/src/model"
-import { MetricContext } from "../MetricContext"
-import { DataContext } from "./DataContext"
+import { MetricContext } from "../contexts/MetricContext"
+import { DataContext } from "../contexts/DataContext"
 
 export function Providers({ children }: { children: React.ReactNode }) {
   const [dataState, setData] = useState<{
