+++ app/src/components/Tooltip.tsx
@@ -3,9 +3,9 @@ import { useMouse } from "react-use"
 import { useMemo, useRef } from "react"
 import styled from "styled-components"
 import { HydratedGitBlobObject } from "../../../parser/src/model"
-import { useOptions } from "../OptionsContext"
+import { useOptions } from "../contexts/OptionsContext"
 import { Spacer } from "./Spacer"
-import { useMetricCaches } from "../MetricContext"
+import { useMetricCaches } from "../contexts/MetricContext"
 import { MetricType } from "../metrics"
 import { dateFormatShort } from "../util"
